- Moves block related svg files to 'images/' to avoid SVGO overrides. - Updates archive link styles for header stories.
